Refactor mail components.
